ANZSCO 2013

Wide table containing all levels of the Australian and New Zealand Standard Classification of Occupations (ANZSCO), Version 1.2, 2013

anzsco2013

Format

A tibble with 11 variables:

anzsco_major_code

Major ANZSCO group codes, e.g. "4"

anzsco_major

Major ANZSCO group names in title case, e.g. "Community and Personal Service Workers"

anzsco_submajor_code

Sub-major ANZSCO group codes, e.g. "45"

anzsco_submajor

Sub-major ANZSCO group names in title case, e.g. "Sports and Personal Service Workers"

anzsco_minor_code

Minor ANZSCO group codes, e.g. "452"

anzsco_minor

Minor ANZSCO group names in title case, e.g. "Sports and Fitness Workers"

anzsco_unit_code

Unit ANZSCO group codes, e.g. "4523"

anzsco_unit

Unit ANZSCO group names in title case, e.g. "Sports Coaches, Instructors and Officials"

anzsco_occupation_code

Occupation ANZSCO group codes, e.g. "452311"

anzsco_occupation

Occupation ANZSCO group names in title case, e.g. "Diving Instructor (Open Water)"

skill_level

Skill level required for occupation, determined by the ABS (1 is highest, 5 is lowest). See https://www.abs.gov.au/ausstats/abs@.nsf/Previousproducts/C4BECE1704987586CA257089001A9181 for details.