Australian Skills Classification (ASC) Core Competencies Descriptions

As described by the National Skills Commission: Core competencies are common to all jobs. They describe a set of non-specialist skills gained in early life and schooling and provide a base to further develop skills and specialties . Popular terms for these include ‘foundation skills’, ‘common skills’, ‘soft skills’, ‘core skills’ and ‘employability skills’. As part of the Australian Skills Classification work, the NSC identified 10 core competencies required for every occupation in Australia. The core competencies align to the definitions of foundation skills typically used in the Australian VET system – specifically, the Employability Skills Framework developed by the Australian Skills Quality Authority, with minor differences recommended by education system experts.

Usage

asc_core_competencies_descriptions

Format

A tibble with:

core_competencies

Character. One of "Numeracy", "Digital engagement", "Teamwork", "Writing", "Reading", "Learning", "Problem solving", "Oral communication", "Planning and organising", "Initiative and innovation". See description

competency_description

Character. See description

score

Integer. See description

proficiency_level

Character. See description

anchor_value

Character. See description

Details

Note: this data includes content of the Australian Skills Classification by the National Skills Commission, Commonwealth of Australia. Used under CC BY 4.0 licence.