Get Australian seasons for given dates

Small helper function that provided a vector with dates, returns the names of corresponding seasons, according to the Australian definition of seasons It can return either "temperate" (4 seasons) or "tropical" (dry/wet) seasons.

Usage

aus_seasons(dates, type = "temperate")

Arguments

dates

vectors with dates. It supports Dates, POSIxt or characters.

type

either "temperate" (default) or "tropical"

Value

vector with season names

Examples

if (FALSE) { # \dontrun{
sample_dates <- c("2022-01-01","2018-10-04","2016-07-20")
# temperate seasons
aus_seasons(sample_dates,"temperate")
# tropical seasons
aus_seasons(sample_dates,"tropical")
} # }