Convert ANZSIC names and abbreviations into the consistent format used by the Australian Bureau of Statistics.

This function enables both exact (default) and fuzzy matching. Under exact matching, if no match is found, NA is returned.

clean_anzsic(
  x,
  fuzzy_match = FALSE,
  max_dist = 0.4,
  method = "jw",
  silent = FALSE
)

Arguments

x
a (character) vector containing ANZSIC titles. Note that clean_anzsic always returns a character vector. If no match is found, then NA is returned.

fuzzy_match
logical; either TRUE which indicates that approximate/fuzzy string matching should be used, or FALSE (the default) which indicates that only exact matches should be used. If FALSE, then if no match is found, then NA is returned.

max_dist
numeric, sets the maximum acceptable distance between your string and the matched string. Default is 0.4. Only relevant when fuzzy_match is TRUE.

method
the method used for approximate/fuzzy string matching. Default is "jw", the Jaro-Winker distance; see `??stringdist-metrics` for more options. Only relevant when fuzzy_match is TRUE.

silent
a boolean value. If FALSE (the default), the function will warn that NAs were returned.

Value

a character vector

See also

Examples

clean_anzsic("Agriculture forestry & Fishing")
#> [1] "Agriculture, Forestry and Fishing"

clean_anzsic("Agriculture, Forestry", fuzzy_match = TRUE)
#> [1] "Agriculture, Forestry and Fishing"