clean_titles is a helper function for the cleaning of titles

clean_titles is a helper function for the cleaning of titles

Usage

clean_titles(
  dictionary = NULL,
  .vector = NULL,
  .fuzzy_match = FALSE,
  .max_dist = 0.4,
  .method = "jw",
  .silent = FALSE
)

Arguments

dictionary

a dictionary created with the make_dictionary function

.vector

a character vector

.fuzzy_match

logical; either TRUE which indicates that approximate/fuzzy string matching should be used, or FALSE (the default) which indicates that only exact matches should be used. If FALSE, then if no match is found, then NA is returned.

.max_dist

numeric, sets the maximum acceptable distance between your string and the matched string. Default is 0.4. Only relevant when fuzzy_match is TRUE.

.method

the method used for approximate/fuzzy string matching. Default is "jw", the Jaro-Winker distance; see `??stringdist-metrics` for more options. Only relevant when fuzzy_match is TRUE.

.silent

a boolean value. If FALSE (the default), the function will warn that NA(s) were returned.

Value

a character vector of cleaned values that have matches in the dictionary.