Identifies if a given date is a public holiday in Australia

Identifies if a given date is a public holiday in Australia

Usage

is_holiday(date, jurisdictions = c())

Arguments

date

Date, POSIXct object or a string that can be parsed with `parsedate::parse_date`

jurisdictions

Vector of state jurisdictions you wish to filter on, defaults to all of Australia

Value

logical(`TRUE` or `FALSE`) vector

Examples

is_holiday('2020-01-01')
#> [1] TRUE
is_holiday('2019-05-27', jurisdictions=c('ACT', 'TAS'))
#> [1] TRUE
h_df <- data.frame(dates = c('2020-01-01', '2020-01-10'))
h_df %>%
  dplyr::mutate(IsHoliday = is_holiday(dates))
#>        dates IsHoliday
#> 1 2020-01-01      TRUE
#> 2 2020-01-10     FALSE