Australian School Terms

Tidy data set of Australian school terms from 1978 to 2024, inclusive. As the data are manually collected by the ABS there is no guarantee of accuracy or comprehensiveness. Missing or obviously incorrect dates are replaced with NA.

Usage

school_terms

Format

A data frame with 1504 rows and 5 variables:

state

The state or territory of the school term

year

The year of the school term

term

The term number of the school term

start

The start date of the school term

end

The end date of the school term

Source

State and Territory governments via the Australian Bureau of Statistics via direct correspondence.