Import SEIFA Data from ABS

The function will download all SEIFA data, for a specified spatial structure, to a temporary excel file and then merge sheets into a single `data.frame`. This `data.frame` also includes the ABS population count for the given spatial structure. For more information on SEIFA indexes go to https://www.abs.gov.au/statistics/people/people-and-communities/socio-economic-indexes-areas-seifa-australia

To learn more about the individual data-sets for each year, please visit their respective pages: 2011 https://www.abs.gov.au/AUSSTATS/abs@.nsf/allprimarymainfeatures/8C5F5BB699A0921CCA258259000BA619 2016 https://www.abs.gov.au/ausstats/abs@.nsf/mf/2033.0.55.001 2021 https://www.abs.gov.au/statistics/people/people-and-communities/socio-economic-indexes-areas-seifa-australia/2021

Usage

get_seifa(
  structure = c("sa1", "sa2", "lga", "postcode", "suburb"),
  data_subclass = c("irsed", "irsead", "ier", "ieo"),
  year = NULL
)

Arguments

structure

character value for the desired spatial area. Must be one of:

sa1 - download size 51.6 MB

sa2 - download size 1.9 MB

lga - download size 660 KB

postcode - download size 2.3 MB

suburb - download size 11.3 MB

data_subclass

character vector matching available SEIFA indexes:

irsed - Index of Relative Socio-economic Disadvantage

irsead - Index of Relative Socio-economic Advantage and Disadvantage

ier - Index of Economic Resources

ieo - Index of Education and Occupation

year

a character string or numeric of the release year of SEIFA object, eg "2016"; 2011.

Value

`data.frame` if file successfully downloaded, else returns `NULL`.

Note

For All ABS SEIFA spreadsheets go to ABS website

Examples

if (FALSE) {
  get_seifa(structure = 'lga', data_subclass = 'irsed', year = 2016)
}