Parse SEIFA index Spreadsheet

This is a helper function used by get_seifa. The function can also be used independently if you have already downloaded one of the SEIFA index score spreadsheets from https://www.abs.gov.au/AUSSTATS/abs@.nsf/DetailsPage/2033.0.55.0012016?OpenDocument

get_seifa_index_sheet(
  filename,
  sheetname,
  structure = c("sa1", "sa2", "lga", "postcode", "suburb")
)

Arguments

filename
character name of temporary file when spreadsheet downloaded

sheetname
character name of the sheet to be imported

structure
character spatial structure of the data to be parsed. The spatial structure is important as the shape of the data in the ABS spreadsheets if different for some structures.

Value

data.frame

Examples

if (FALSE) {

  get_seifa_index_sheet('downloaded_filename.xls', sheetname = 'Table 2', structure = 'lga')
}